Moved routes into seperate file
